Fix bugs discovered during code review
