+++ frontend/src/pages/audio/collection.vue
@@ -4,22 +4,29 @@
       v-if="collectionParams"
       search-term=""
       :is-fetching="isFetching"
-      :results="results"
+      :results="{ type: 'audio', items: media }"
       :collection-label="collectionLabel"
       :collection-params="collectionParams"
-      @load-more="handleLoadMore"
+      @load-more="loadMore"
     />
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent, useFetch, useMeta } from "@nuxtjs/composition-api"
-import { computed } from "vue"
+import {
+  defineComponent,
+  useFetch,
+  useMeta,
+  useRoute,
+} from "@nuxtjs/composition-api"
+import { computed, ref, watch } from "vue"
 
 import { collectionMiddleware } from "~/middleware/collection"
 import { useSearchStore } from "~/stores/search"
-import { useCollection } from "~/composables/use-collection"
-import { AUDIO } from "~/constants/media"
+
+import { useMediaStore } from "~/stores/media"
+import type { AudioDetail } from "~/types/media"
+import { useI18n } from "~/composables/use-i18n"
 
 import VCollectionResults from "~/components/VSearchResultsGrid/VCollectionResults.vue"
 
@@ -29,36 +36,63 @@ export default defineComponent({
   layout: "content-layout",
   middleware: collectionMiddleware,
   setup() {
+    const mediaStore = useMediaStore()
     const searchStore = useSearchStore()
 
     const collectionParams = computed(() => searchStore.collectionParams)
+    const isFetching = computed(() => mediaStore.fetchState.isFetching)
 
-    const {
-      results,
-      creatorUrl,
-      fetchMedia,
-      handleLoadMore,
-      collectionLabel,
-      isFetching,
-    } = useCollection({
-      mediaType: AUDIO,
+    const media = ref<AudioDetail[]>([])
+    const creatorUrl = ref<string>()
+
+    const i18n = useI18n()
+
+    const collectionLabel = computed(() => {
+      if (!collectionParams.value) {
+        return ""
+      }
+      const { collection, ...params } = collectionParams.value
+      return i18n
+        .t(`collection.label.${collection}.audio`, { ...params })
+        .toString()
+    })
+
+    const fetchMedia = async (
+      { shouldPersistMedia }: { shouldPersistMedia: boolean } = {
+        shouldPersistMedia: false,
+      }
+    ) => {
+      media.value = (await mediaStore.fetchMedia({
+        shouldPersistMedia,
+      })) as (typeof media)["value"]
+      creatorUrl.value =
+        media.value.length > 0 ? media.value[0].creator_url : undefined
+    }
+    const loadMore = () => {
+      fetchMedia({ shouldPersistMedia: true })
+    }
+
+    useMeta({
+      meta: [{ hid: "robots", name: "robots", content: "all" }],
     })
 
     useFetch(async () => {
       await fetchMedia()
     })
 
-    useMeta({
-      meta: [{ hid: "robots", name: "robots", content: "all" }],
+    // Fetch media when the route changes, e.g. when the user navigates from
+    // a creator collection page to a source collection page.
+    const route = useRoute()
+    watch(route, () => {
+      fetchMedia()
     })
 
     return {
-      results,
       collectionParams,
       isFetching,
-      creatorUrl,
+      media,
       collectionLabel,
-      handleLoadMore,
+      loadMore,
     }
   },
   head: {},
